sys/ztimer: Use uint32_t for freq in periph_timer
Now that timer_init() of periph_timer uses uint32_t for the frequency, ztimer
should do the same.

+++ sys/include/ztimer/periph_timer.h
@@ -54,7 +54,7 @@ typedef struct {
  * @param[in]   max_val maximum value this timer supports
  */
 void ztimer_periph_timer_init(ztimer_periph_timer_t *clock, tim_t dev,
-                              unsigned long freq, uint32_t max_val);
+                              uint32_t freq, uint32_t max_val);
 
 #ifdef __cplusplus
 }
